+++ lib/lisp.rb
@@ -1,5 +1,9 @@
 require "lisp/version"
 require "lisp/lisp"
+require "lisp/parser"
+require "lisp/evaluator"
+require "lisp/environment"
+require "lisp/procedure"
 
 module Lisp
 end
