+++ lib/lisp/evaluator.rb
@@ -1,0 +1,39 @@
+module Lisp
+  class Evaluator
+    def initialize
+      @global_env = Environment.standard_env
+    end
+
+    def lisp_eval(x, env = @global_env)
+      # Evaluate an expression in an environment."
+      if x.is_a? Symbol      # variable reference
+        return env.find(x)[x]
+      elsif !x.is_a? Array   # constant literal
+        return x
+      elsif x[0] == :quote          # (quote exp)
+        return x[1..-1]
+      elsif x[0] == :if             # (if test conseq alt)
+        _, test, conseq, alt = x
+        exp = lisp_eval(test, env) ? conseq : alt
+        return lisp_eval(exp, env)
+      elsif x[0] == :define         # (define var exp)
+        var = x[1]
+        exp = x[2]
+        env[var] = lisp_eval(exp, env)
+      elsif x[0] == :exit
+        :exit
+      elsif x[0] == :quit
+        :exit
+      elsif x[0] == :lambda         # procedure
+        # (_, parms, body) = x
+        parms = x[1]
+        body = x[2]
+        return Procedure.new(self, parms, body, env)
+      else                          # (proc arg...)
+        p = lisp_eval(x[0], env)
+        args = x[1..-1].map { |arg| lisp_eval(arg, env) }
+        return p.call(*args)
+      end
+    end
+  end
+end
